pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-anomaly"
version = "0.1.0"
description = "Retail store scan-quality scoring: LFM + GBDT + Conditional VAE"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Gary Huang" }]

dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "torch>=2.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "mlflow>=2.10",
    "shap>=0.44",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "lightgbm>=4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "httpx>=0.26",       # for FastAPI TestClient
    "ruff>=0.2",
    "mypy>=1.8",
]

[project.scripts]
retail-score   = "retail_anomaly.cli:score"
retail-generate = "retail_anomaly.cli:generate"

[tool.setuptools.packages.find]
where = ["."]
include = ["retail_anomaly*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "--cov=retail_anomaly --cov-report=term-missing -q"
